change database model
